refactor: 메뉴 정보 jpa Long 타입 Integer로 변경

+++ src/main/java/com/oopsw/seongsubean/cafe/repository/MenuInfoRepository.java
@@ -3,6 +3,6 @@ package com.oopsw.seongsubean.cafe.repository;
 import com.oopsw.seongsubean.cafe.domain.MenuInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuInfoRepository extends JpaRepository<MenuInfo, Long> {
+public interface MenuInfoRepository extends JpaRepository<MenuInfo, Integer> {
 
 }
